+++ lib/simple_form/action_view_extensions/form_helper.rb
@@ -10,7 +10,7 @@ module SimpleForm
     #
     module FormHelper
 
-      def simple_form_for(record, options={}, &block)
+      def simple_form_for(record, options = {}, &block)
         options[:builder] ||= SimpleForm::FormBuilder
         options[:html] ||= {}
         unless options[:html].key?(:novalidate)