+++ lib/simple_form/components/readonly.rb
@@ -2,7 +2,7 @@ module SimpleForm
   module Components
     # Needs to be enabled in order to do automatic lookups.
     module Readonly
-      def readonly
+      def readonly(wrapper_options = nil)
         if readonly_attribute? && !has_readonly?
           input_html_options[:readonly] ||= true
           input_html_classes << :readonly