+++ lib/simple_form/inputs/date_time_input.rb
@@ -1,11 +1,13 @@
 module SimpleForm
   module Inputs
     class DateTimeInput < Base
-      def input
+      def input(wrapper_options = nil)
+        merged_input_options = merge_wrapper_options(input_html_options, wrapper_options)
+
         if use_html5_inputs?
-          @builder.send(:"#{input_type}_field", attribute_name, input_html_options)
+          @builder.send(:"#{input_type}_field", attribute_name, merged_input_options)
         else
-          @builder.send(:"#{input_type}_select", attribute_name, input_options, input_html_options)
+          @builder.send(:"#{input_type}_select", attribute_name, input_options, merged_input_options)
         end
       end