+++ lib/simple_form/inputs/grouped_collection_select_input.rb
@@ -1,11 +1,14 @@
 module SimpleForm
   module Inputs
     class GroupedCollectionSelectInput < CollectionInput
-      def input
+      def input(wrapper_options = nil)
         label_method, value_method = detect_collection_methods
+
+        merged_input_options = merge_wrapper_options(input_html_options, wrapper_options)
+
         @builder.grouped_collection_select(attribute_name, grouped_collection,
                       group_method, group_label_method, value_method, label_method,
-                      input_options, input_html_options)
+                      input_options, merged_input_options)
       end
 
       private