+++ lib/simple_form/inputs/password_input.rb
@@ -3,8 +3,10 @@ module SimpleForm
     class PasswordInput < Base
       enable :placeholder, :maxlength
 
-      def input
-        @builder.password_field(attribute_name, input_html_options)
+      def input(wrapper_options = nil)
+        merged_input_options = merge_wrapper_options(input_html_options, wrapper_options)
+
+        @builder.password_field(attribute_name, merged_input_options)
       end
     end
   end