+++ lib/simple_form/inputs/priority_input.rb
@@ -1,9 +1,11 @@
 module SimpleForm
   module Inputs
     class PriorityInput < CollectionSelectInput
-      def input
+      def input(wrapper_options = nil)
+        merged_input_options = merge_wrapper_options(input_html_options, wrapper_options)
+
         @builder.send(:"#{input_type}_select", attribute_name, input_priority,
-                      input_options, input_html_options)
+                      input_options, merged_input_options)
       end
 
       def input_priority