+++ lib/simple_form/inputs/range_input.rb
@@ -1,7 +1,7 @@
 module SimpleForm
   module Inputs
     class RangeInput < NumericInput
-      def input
+      def input(wrapper_options = nil)
         if html5?
           input_html_options[:type] ||= "range"
           input_html_options[:step] ||= 1