+++ lib/simple_form/wrappers/builder.rb
@@ -45,20 +45,20 @@ module SimpleForm
         @components = []
       end
 
-      def use(name, options=nil, &block)
+      def use(name, options = {}, &block)
         if options && wrapper = options[:wrap_with]
-          @components << Single.new(name, wrapper)
+          @components << Single.new(name, wrapper, options.except(:wrap_with))
         else
-          @components << name
+          @components << Leaf.new(name, options)
         end
       end
 
-      def optional(name, options=nil, &block)
+      def optional(name, options = {}, &block)
         @options[name] = false
         use(name, options, &block)
       end
 
-      def wrapper(name, options=nil)
+      def wrapper(name, options = nil)
         if block_given?
           name, options = nil, name if name.is_a?(Hash)
           builder = self.class.new(@options)