+++ lib/simple_form/wrappers/many.rb
@@ -1,7 +1,7 @@
 module SimpleForm
   module Wrappers
     # A wrapper is an object that holds several components and render them.
-    # A component may either be a symbol or any object that responds to `render`.
+    # A component may be any object that responds to `render`.
     # This API allows inputs/components to be easily wrapped, removing the
     # need to modify the code only to wrap input in an extra tag.
     #
@@ -10,9 +10,8 @@ module SimpleForm
     # on demand on input generation.
     class Many
       attr_reader :namespace, :defaults, :components
-      alias :to_sym :namespace
 
-      def initialize(namespace, components, defaults={})
+      def initialize(namespace, components, defaults = {})
         @namespace  = namespace
         @components = components
         @defaults   = defaults
@@ -25,8 +24,8 @@ module SimpleForm
         options = input.options
 
         components.each do |component|
-          next if options[component] == false
-          rendered = component.respond_to?(:render) ? component.render(input) : input.send(component)
+          next if options[component.namespace] == false
+          rendered = component.render(input)
           content.safe_concat rendered.to_s if rendered
         end