+++ lib/simple_form/wrappers/single.rb
@@ -2,14 +2,16 @@ module SimpleForm
   module Wrappers
     # `Single` is an optimization for a wrapper that has only one component.
     class Single < Many
-      def initialize(name, options={})
-        super(name, [name], options)
+      def initialize(name, wrapper_options = {}, options = {})
+        @component = Leaf.new(name, options)
+
+        super(name, [@component], wrapper_options)
       end
 
       def render(input)
         options = input.options
         if options[namespace] != false
-          content = input.send(namespace)
+          content = @component.render(input)
           wrap(input, options, content) if content
         end
       end